test: Move fixtures from /packages/messageformat/src/ to /test/
